[feat]拆分AbstractWebApiDependencyRegistrar.cs的代码到多个文件。
