+++ src/ServerApi/Services/Shared/Adnc.Shared.WebApi/Registrar/AbstractWebApiDependencyRegistrar.MiniProfiler.cs
@@ -1,0 +1,12 @@
+﻿namespace Adnc.Shared.WebApi.Registrar;
+
+public abstract partial class AbstractWebApiDependencyRegistrar
+{
+    /// <summary>
+    /// 注册 MiniProfiler 组件
+    /// </summary>
+    protected virtual void AddMiniProfiler() =>
+        Services
+        .AddMiniProfiler(options => options.RouteBasePath = $"/{ServiceInfo.ShortName}/profiler")
+        .AddEntityFramework();
+}
